requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.3
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
autopep8==1.4.4
awsebcli==3.16.0
beautifulsoup4==4.9.0
blessed==1.16.1
blinker==1.4
boto3==1.10.41
botocore==1.13.41
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==6.7
#cloud-init==20.4
colorama==0.3.9
#command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
distlib==0.3.1
#distro-info===0.18ubuntu0.18.04.1
Django==3.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-cors-headers==3.3.0
django-fernet-fields==0.6
django-filter==2.2.0
django-nested-admin==3.3.3
django-phone-field==1.8.0
django-rest-swagger==2.2.0
django-storages==1.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.17.1
filelock==3.0.12
future==0.16.0
httplib2==0.9.2
hyperlink==17.3.1
#idna==2.9
importlib-metadata==3.4.0
importlib-resources==5.1.0
incremental==16.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.10
jmespath==0.9.4
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
#language-selector==0.1
#MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
openapi-codec==1.3.2
packaging==20.9
#PAM==0.4.2
pathspec==0.5.9
# psycopg2==2.8.4
pyarmor==6.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
#pygobject==3.26.1
PyJWT==1.7.1
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyserial==3.4
#python-apt==1.6.5+ubuntu0.5
python-crontab==2.4.2
python-dateutil==2.8.0
python-debian==0.1.32
python-monkey-business==1.0.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.13
#requests==2.23.0
requests-unixsocket==0.1.5
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.2.1
SecretStorage==2.3.1
semantic-version==2.5.0
service-identity==16.0.0
simplejson==3.17.2
six==1.11.0
soupsieve==2.0
sqlparse==0.3.0
#ssh-import-id==5.7
#systemd-python==234
termcolor==1.1.0
texttable==0.9.1
Twisted==17.9.0
typing-extensions==3.7.4.3
tzlocal==2.1
#ufw==0.36
#unattended-upgrades==0.1
uritemplate==3.0.1
#urllib3==1.25.8
virtualenv==20.4.2
wcwidth==0.1.7
websocket-client==0.56.0
zipp==3.4.0
#zope.interface==4.3.2
